Refactored the fine-tuning code into model.py
